requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymupdf
pypdf
faiss-cpu
sentence-transformers
scikit-learn
numpy
pandas
pydantic
requests
streamlit
google-generativeai
supabase
postgrest
PyJWT
python-multipart
fastapi
uvicorn[standard]
python-jose[cryptography]
httpx
python-dotenv

